Less warnings during compilation

+++ libcheetah/src/saveFrame.cpp
@@ -381,7 +381,7 @@ void writeHDF5(cEventData *info, cGlobal *global){
 		double *peak_info = (double *) calloc(4*size[0], sizeof(double));
 		
 		// Set all unused peaks to 0
-		for(long i=0; i< 4*size[0]; i++) {
+		for(uint i=0; i< 4*size[0]; i++) {
 			peak_info[i] = 0;
 		}
